factor out the node-free detector
