+++ apriltag_detector_mit/launch/composable.launch.py
@@ -1,5 +1,5 @@
 # -----------------------------------------------------------------------------
-# Copyright 2022 Bernd Pfrommer <bernd.pfrommer@gmail.com>
+# Copyright 2024 Bernd Pfrommer <bernd.pfrommer@gmail.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
