+++ apriltag_detector_mit/src/node.cpp
@@ -13,15 +13,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <apriltag_detector_mit/apriltag_detector.hpp>
+#include <apriltag_detector_mit/component.hpp>
 #include <memory>
 #include <rclcpp/rclcpp.hpp>
 
 int main(int argc, char ** argv)
 {
   rclcpp::init(argc, argv);
-  auto node = std::make_shared<apriltag_detector_mit::ApriltagDetector>(
-    rclcpp::NodeOptions());
+  auto node =
+    std::make_shared<apriltag_detector_mit::Component>(rclcpp::NodeOptions());
   rclcpp::spin(node);  // should not return
   rclcpp::shutdown();
   return 0;
